benchmarks/requirements.txt
redshift-connector>=2.1.0
pyarrow>=14.0.0
boto3>=1.34.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.8.0
